Authorisation with policies

+++ app/Http/Controllers/ThreadsController.php
@@ -90,16 +90,14 @@ class ThreadsController extends Controller
     {
         //
     }
-    
+
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->replies()->delete();
         $thread->delete();
-        if(request()->wantsJson()) {
-            return response([], 204);
-        } else {
-            return redirect('/threads/');
-        }
+        return redirect('/threads/');
     }
 
     private function getThreads(ThreadFilters $filters, Channel $channel)
